experiments/weather-pets/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "codex-pet-weather"
version = "0.1.0"
description = "Weatherize Codex custom pet spritesheets without patching Codex."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Codex Pet Weather Contributors" }]
dependencies = ["Pillow>=10"]

[project.scripts]
codex-pet-weather = "codex_pet_weather.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
